Removed unnecessary unboxing.

+++ src/main/java/com/jhdit/algorithms/frequency/MostFrequentNumberUsingMap.java
@@ -58,7 +58,7 @@ public class MostFrequentNumberUsingMap implements MostFrequentNumber {
         return values.stream().max(new Comparator<Integer>() {
             @Override
             public int compare(Integer o1, Integer o2) {
-                if (o1.intValue() > o2.intValue())  {
+                if (o1 > o2)  {
                     return 1;
                 }
                 return -1;
